Generalized Makefile
While I'm not completely sure I've achieved what I was looking for, I believe I've adjusted my Makefile to be much more generalized and clean.
For the scope of this program, I no longer need to edit this when adding new implementation or changing file names.
Beyond the scope of this project, I believe I am able to reuse this Makefile for other programs with minimal to no changes.
There are a few improvements I would like to make to this at some point, such as add functionality so that the Makefile doesn't need to be in the root directory and so that the source files gathered can be more customizable than "everything within the root directory", but I'm finding it to be good enough as is.