Began Generation Methods
MazeGame has been renamed to MapInterface
MapInterface now has 4 extra functions to construct the Map:
* TwoRooms - The already implemented hardcoded version
* Procedural
* Binary Tree
* Kruskal
The latter 3 currently only build a single room surrounded by walls, but will be the sites of algorithm implementation to generate maps in different ways. Additionally, an enum will be introduced to be passed into MapInterface::BuildMaze to select one of these functions to use.