MapGenerator made + small fixes/cleanup
The code that generates a maze has been pulled out of MapInterface and into its own class called MapGenerator, with a function for each type of generation.
With this, Main and MapInterface are now solely for a complete program that can be run in the terminal, and will be excluded when this system is imported into Unreal Engine, where Blueprints will functionally replace MapInterface.